@@ -6,22 +6,22 @@
     "start": "NODE_ENV=production node server.js"
   },
   "dependencies": {
-    "next": "^3.2.2",
-    "next-routes": "^1.0.40",
+    "next": "^4.1.4",
+    "next-routes": "^1.1.0",
     "nprogress": "^0.2.0",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-ga": "^2.2.0",
-    "react-images": "^0.5.6",
+    "react": "^16.1.1",
+    "react-dom": "^16.1.1",
+    "react-ga": "^2.3.5",
+    "react-images": "^0.5.8",
     "react-tinymce-input": "^2.1.0"
   },
   "devDependencies": {
-    "babel-eslint": "^8.0.0",
-    "eslint": "^4.7.0",
-    "eslint-config-airbnb": "^15.1.0",
-    "eslint-plugin-import": "^2.7.0",
+    "babel-eslint": "^8.0.2",
+    "eslint": "^4.11.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.3.0"
+    "eslint-plugin-react": "^7.5.0"
   },
   "eslintConfig": {
     "parser": "babel-eslint",
